refactoring/팀원과 파일 구조 맞춤
