--- src/App.js
@@ -1,5 +1,0 @@
-function App() {
-
-}
-
-export default App;
