+++ src/components/Gnb.js
@@ -1,5 +1,4 @@
-import logo from './assets/logo.jpg';
-
+import logo from '../assets/logo.jpg'
 function Gnb(){
     return(
         <>
