+++ src/index.js
@@ -1,11 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import GroupPage from './GroupPage.js';
+import GroupPage from './pages/GroupPage.js';
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
-    <GroupPage />
-    
+    <GroupPage /> 
   </React.StrictMode>
 );
