+++ src/pages/GroupPage.js
@@ -1,9 +1,9 @@
-import Card from './CardMemory.js';
-import Info from './Info.js';
-import Gnb from './Gnb.js';
-import Button from './Button.js';
-import Tab from './Tab.js';
-import Search from './Search.js';
+import Card from '../components/CardMemory.js';
+import Info from '../components/Info.js';
+import Gnb from '../components/Gnb.js';
+import Button from '../components/Button.js';
+import Tab from '../components/Tab.js';
+import Search from '../components/Search.js';
 const style={
     margin:'20px',
     display: 'flex',
